Move change password functions from modules to utils
